// src/app/interfaces/delivery-record.ts
export interface DeliveryRecord {
    Id: number;
    ReferenceNumber: string;
    TrackNumber: string;
    PriceName: string;
    PriceCode?: string;
    Weight: string;
    Amount: string;
    Date: string;
    Selected: boolean;
    ModeOfTransportName: string;
    ProductContent: string;
    ChargeableWeight: string;
    Piece: number;
    CountryName: string;
    CustomerName?: string;
    PalletNo?: string;
    TransportIdDocumentId?: number;
    IsRemote?: boolean;
    CountryNo?: string;
    CountryNameCn?: string;
    CountryNameCN?: string;
    CountryNameZh?: string;
    CountryZh?: string;
    CountryChinese?: string;
    Currency?: string;
    IsReturn?: boolean;
    Remark?: string;
    CurrentActivity?: string;
    ConsigneeName?: string;
    ConsigneeCompany?: string;
    AttributeNames?: string | null;
    VolumeWeightOf5000?: number;
    RgdDate?: string;
    ActualWeight?: number;
    VolumeWeight?: number;
    IsAmazon?: boolean;
    PackagingOptimizationStatus?: number;
    SalesmanName?: string;
    MerchandiserName?: string;
    FBAStatus?: string;
}